requirements.txt
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
oauthlib==2.1.0
praw==6.0.0
prawcore==1.0.0
requests==2.21.0
requests-oauthlib==1.0.0
tqdm==4.29.1